backend/requirements.txt
fastapi
uvicorn
google-generativeai
qdrant-client
aiosqlite
python-dotenv